Update SConstruct for changes to ALTA
It seems that at least since v0.2 of ALTA, the directory structure has changed, and the build file of this plugin needed to be changed to account for that.

Also, it might be worthwhile to add an error message when a user hasn't supplied a binary file, rather then crashing with a segfault, or at least make a mention of this in the README. When I first tried to use this, I was helping a friend to compile it, so I wasn't fully aware of what it does. I assumed that it searches for pre-packaged files in the companion alta directory, but it doesn't. The segfault confused me for a bit of time and I at first thought that we wouldn't be able to do anything worthwhile.

Best,
Rob